Move "noise" ring generation to NoisePing.cs
