+++ Assets/Prefabs/Interactables/PingRing.prefab.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: c6cff22645e904a40bb95ed17be87a89
+guid: e14dd387ab2fad4439dc7b3cc92aa004
 PrefabImporter:
   externalObjects: {}
   userData:
